清单添加了按店删除的功能

+++ com.yrtech.easyPhotoAPI/com.yrtech.InventoryAPI/DTO/Answer/AnswerPhotoDto.cs
@@ -8,7 +8,7 @@ namespace com.yrtech.InventoryAPI.DTO
     [Serializable]
     public class AnswerPhotoDto
     {
-        public int AnswerId { get; set; }
+        public long AnswerId { get; set; }
         public int PhotoId { get; set; }
         public int ProjectId { get; set; }
         public string CheckCode { get; set; }
